+++ src/main/java/lotto/dto/input/LastWinningData.java
@@ -1,0 +1,4 @@
+package lotto.dto.input;
+
+public class LastWinningData {
+}
